Time: 2 ms (61.06%), Space: 6.3 MB (27.84%) - LeetHub

+++ 0096-unique-binary-search-trees/0096-unique-binary-search-trees.cpp
@@ -5,6 +5,6 @@ public:
         for (int i = 0; i < n; ++i) {
             result = result * 2 * (2 * i + 1) / (i + 2);
         }
-        return static_cast<int>(result); 
+        return result;
     }
 };
